backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-opencv==4.8.1.78
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
rembg==2.0.50
onnxruntime==1.16.0
mediapipe==0.10.7
reportlab==4.0.4
fpdf2==2.7.6
aiofiles==23.2.1
pydantic==2.4.2
starlette==0.27.0
typing-extensions==4.8.0
requests==2.31.0
scikit-image==0.21.0
matplotlib==3.7.2